Move create project to separate page, edit support and refactoring (#321)
* moved create project to page, added roles and edit support

* show success modal after project is created

* fmt

* rebased and added reusable view container

* bug fixes

* add logs

* css update

* update login container css
