+++ apps/new/widget/lib/projects.jsx
@@ -75,6 +75,27 @@ const fetchProjects = (props) => {
   return processData(data, type);
 };
 
+const getProjectMeta = (id) => {
+  if (!id) {
+    throw new Error("Invalid project ID");
+  }
+
+  const data = Social.get(id, "final");
+
+  if (!data) {
+    console.log("Failed to fetch project data");
+  }
+
+  try {
+    const pj = JSON.parse(data);
+    return pj;
+  } catch (error) {
+    console.error("Error parsing project data:", error);
+    return null;
+  }
+};
+
 return {
   fetchProjects,
+  getProjectMeta,
 };
