+++ apps/new/widget/page/projects/SearchBar.jsx
